ads-weekly-report-kit/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ads-weekly-report-kit"
version = "0.1.0"
description = "Meta Ads CSV → 10-min weekly report (HTML + Excel)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "adswk" }]
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "matplotlib>=3.7",
  "openpyxl>=3.1",
  "jinja2>=3.1",
  "PyYAML>=6.0",
]

[project.scripts]
adswk = "adswk.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

